+++ cedar-policy/src/integration_testing.rs
@@ -33,8 +33,8 @@
 #![allow(clippy::expect_used)]
 
 use crate::{
-    frontend::is_authorized::InterfaceResponse, Authorizer, Context, Decision, Entities, EntityUid,
-    Policy, PolicyId, PolicySet, Request, Schema, ValidationMode, Validator,
+    frontend::is_authorized::InterfaceResponse, AuthorizationError, Authorizer, Context, Decision,
+    Entities, EntityUid, Policy, PolicyId, PolicySet, Request, Schema, ValidationMode, Validator,
 };
 use cedar_policy_core::jsonvalue::JsonValueWithNoDuplicateKeys;
 use serde::{Deserialize, Serialize};
@@ -376,8 +376,12 @@ pub fn perform_integration_test_from_json_custom(
                 &json_request.desc
             );
             // check errors
-            let errors: HashSet<PolicyId> =
-                response.diagnostics().error_policy_ids().cloned().collect();
+            let errors: HashSet<PolicyId> = response
+                .diagnostics()
+                .errors()
+                .map(AuthorizationError::id)
+                .cloned()
+                .collect();
             assert_eq!(
                 errors,
                 json_request.errors.into_iter().collect(),